Move BuildPath to com.google.cloud.tools.eclipse.appengine.libraries
