+++ plugins/com.google.cloud.tools.eclipse.appengine.libraries/META-INF/MANIFEST.MF
@@ -26,6 +26,8 @@ Import-Package: com.google.cloud.tools.eclipse.util,
  org.eclipse.core.resources,
  org.eclipse.core.runtime;version="3.5.0",
  org.eclipse.core.runtime.jobs,
+ org.eclipse.e4.core.contexts,
+ org.eclipse.e4.core.di,
  org.eclipse.e4.core.di.annotations,
  org.eclipse.jdt.core,
  org.eclipse.jst.j2ee.classpathdep,
