+++ plugins/com.google.cloud.tools.eclipse.appengine.newproject/META-INF/MANIFEST.MF
@@ -35,8 +35,6 @@ Import-Package: com.google.cloud.tools.eclipse.appengine.facets,
  org.eclipse.core.runtime;version="3.5.0",
  org.eclipse.core.runtime.jobs,
  org.eclipse.core.runtime.preferences;version="3.3.0",
- org.eclipse.e4.core.contexts,
- org.eclipse.e4.core.di,
  org.eclipse.jdt.core,
  org.eclipse.jdt.junit,
  org.eclipse.jface.databinding.swt,
